Put WebResult into a seperate file
