+++ src/common/announce_multiplayer_room.h
@@ -9,23 +9,7 @@
 #include <string>
 #include <vector>
 #include "common/common_types.h"
-
-namespace Common {
-struct WebResult {
-    enum class Code : u32 {
-        Success,
-        InvalidURL,
-        CredentialsMissing,
-        LibError,
-        HttpError,
-        WrongContent,
-        NoWebservice,
-    };
-    Code result_code;
-    std::string result_string;
-    std::string returned_data;
-};
-} // namespace Common
+#include "common/web_result.h"
 
 namespace AnnounceMultiplayerRoom {
